Create Netflix Movie Analysis With Python
Starting as a DVD rental service in 1997, Netflix has become the largest entertainment and media company by market capitalization with over 200 million subscribers as of January 2021. With an extensive collection of movies and series, Netflix provides a great opportunity to dive into the entertainment industry and showcase my data manipulation skills using Python. 
Nb: This is a project on data camp